+++ src/lte/examples/d2d-examples/lte-sl-out-of-covrg-discovery-collision.cc
@@ -231,8 +231,8 @@ int main (int argc, char *argv[])
 
 
   NS_LOG_INFO ("Configuring discovery applications");
-   std::map<Ptr<NetDevice>, std::list<uint64_t> > announceApps;
-   std::map<Ptr<NetDevice>, std::list<uint64_t> > monitorApps;
+   std::map<Ptr<NetDevice>, std::list<uint32_t> > announceApps;
+   std::map<Ptr<NetDevice>, std::list<uint32_t> > monitorApps;
    
    for (uint32_t i = 1; i <= ueNodes.GetN (); ++i)
    {
@@ -252,9 +252,9 @@ int main (int argc, char *argv[])
    for (auto itAnnounceApps : announceApps)
    {
       Ptr<LteUeNetDevice> ueNetDevice = DynamicCast<LteUeNetDevice> (itAnnounceApps.first);
-      std::list<uint64_t> apps = itAnnounceApps.second;
+      std::list<uint32_t> apps = itAnnounceApps.second;
       std::cout << "Scheduling " << apps.size () << " announce apps for UE with IMSI = " << ueNetDevice->GetImsi () << std::endl;
-      std::list<uint64_t>::iterator itAppList;
+      std::list<uint32_t>::iterator itAppList;
       for (auto itAppList : apps)
       {
          std::cout << "Announcing App code = " << itAppList << std::endl;
@@ -266,9 +266,9 @@ int main (int argc, char *argv[])
    for (auto itMonitorApps : monitorApps)
    {
       Ptr<LteUeNetDevice> ueNetDevice = DynamicCast<LteUeNetDevice> (itMonitorApps.first);
-      std::list<uint64_t> apps = itMonitorApps.second;
+      std::list<uint32_t> apps = itMonitorApps.second;
       std::cout << "Scheduling " << apps.size () << " monitor apps for UE with IMSI = " << ueNetDevice->GetImsi () << std::endl;
-      std::list<uint64_t>::iterator itAppList;
+      std::list<uint32_t>::iterator itAppList;
       for (auto itAppList : apps)
       {
          std::cout << "Monitoring App code = " << itAppList << std::endl;
